+++ nri-kafka/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.0.5
+
+- Added `StatsCallback` support to `Kafka` and `Kafka.Worker`.
+- Added `STATISTICS_INTERVAL_MS` environment variable to `Kafka` and `Kafka.Worker`.
+
 # 0.1.0.4
 
 - Added new `ElsewhereButToKafkaAsWell` mode to `CommitOffsets`, which commits offsets to Kafka once the external Offset storage has been updated. Kafka commits are performed only to keep Kafka informed about consumer lag.